requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
django-filter==23.5
geopy==2.4.1